Added cuda keywords to model, data instantiation.
Should I rewrite the DataSet so that image loading occurs on an
as-needed basis. I worry that loading all of the images into memory at
once is unnecessary. But then would it take too long if, for each batch,
the images needed to be loaded into mem? Probably better as it is. Still
I wonder.